NF: factorize onOptionsItemSelected
